Refactored the main project to remove unnecessary dependencies. Also removed some visual studio detritus and removed unused using statements.
