+++ MarkovSharp.Midi/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Sanford.Multimedia.Midi" version="6.1.1" targetFramework="net461" />
+</packages>
