+++ MarkovSharp.sln
@@ -1,13 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.6
+VisualStudioVersion = 15.0.27004.2006
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarkovSharp", "MarkovSharp\MarkovSharp.csproj", "{D3003B4E-0007-43EB-B350-2410286C6DB5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarkovSharp.Tests", "MarkovSharp.Tests\MarkovSharp.Tests.csproj", "{C484CBE3-CD29-457E-B599-893C878CE521}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarkovSite", "MarkovSite\MarkovSite.csproj", "{E12812E6-6569-45C8-AD94-D077D2E795F7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarkovSharp.Midi", "MarkovSharp.Midi\MarkovSharp.Midi.csproj", "{85C338F8-6F33-48E1-B1AA-300D8745CE98}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,8 +29,15 @@ Global
 		{E12812E6-6569-45C8-AD94-D077D2E795F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E12812E6-6569-45C8-AD94-D077D2E795F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E12812E6-6569-45C8-AD94-D077D2E795F7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{85C338F8-6F33-48E1-B1AA-300D8745CE98}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85C338F8-6F33-48E1-B1AA-300D8745CE98}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85C338F8-6F33-48E1-B1AA-300D8745CE98}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85C338F8-6F33-48E1-B1AA-300D8745CE98}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A22B229A-E67B-4991-B31E-13B29B300B44}
 	EndGlobalSection
 EndGlobal
