--- MarkovSharp/log4net.config
@@ -1,65 +1,0 @@
-<configuration>
-  <configSections>
-    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net"/>
-  </configSections>
-  <log4net>
-    <appender name="Log4ViewRollingFile" type="log4net.Appender.RollingFileAppender">
-      <file value="Logs/Markov.log" />
-      <appendToFile value="true" />
-      <maximumFileSize value="5MB" />
-      <maxSizeRollBackups value="5" />
-      <layout type="log4net.Layout.XmlLayout"/>
-    </appender>
-
-    <appender name="ColoredConsole" type="log4net.Appender.ColoredConsoleAppender">
-      <mapping>
-        <level value="FATAL" />
-        <foreColor value="White" />
-        <backColor value="Red, HighIntensity" />
-      </mapping>
-      <mapping>
-        <level value="ERROR" />
-        <foreColor value="Red, HighIntensity" />
-      </mapping>
-      <mapping>
-        <level value="WARN" />
-        <foreColor value="Yellow, HighIntensity" />
-      </mapping>
-      <mapping>
-        <level value="INFO" />
-        <foreColor value="Green" />
-      </mapping>
-      <mapping>
-        <level value="DEBUG" />
-        <foreColor value="White" />
-      </mapping>
-      <mapping>
-        <level value="VERBOSE" />
-        <foreColor value="Green, HighIntensity" />
-      </mapping>
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%message%newline" />
-      </layout>
-      <filter type="log4net.Filter.LevelRangeFilter">
-        <levelMin value="DEBUG" />
-        <levelMax value="FATAL" />
-      </filter>
-    </appender>
-    <appender name="Console" type="log4net.Appender.ConsoleAppender">
-      <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date{HH:mm:ss} %-5level [%method]: %message%newline" />
-      </layout>
-      <filter type="log4net.Filter.LevelRangeFilter">
-        <levelMin value="DEBUG" />
-        <levelMax value="FATAL" />
-      </filter>
-    </appender>
-    
-    <root>
-      <level value="DEBUG" />
-      <appender-ref ref="ColoredConsole" />
-      <appender-ref ref="Console" />
-      <appender-ref ref="Log4ViewRollingFile" />
-    </root>
-  </log4net>
-</configuration>
