--- MarkovSharp/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="log4net" version="2.0.5" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net452" />
-  <package id="Sanford.Multimedia.Midi" version="6.1.1" targetFramework="net452" />
-</packages>
