+++ MarkovSite/App_Start/FilterConfig.cs
@@ -1,5 +1,4 @@
-﻿using System.Web;
-using System.Web.Mvc;
+﻿using System.Web.Mvc;
 
 namespace MarkovSite
 {
