+++ MarkovSite/Controllers/HomeController.cs
@@ -2,10 +2,8 @@
 using MarkovSite.Models;
 using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Web;
 using System.Web.Mvc;
 
 namespace MarkovSite.Controllers
