+++ MarkovSite/Models/GetPredictionsRequest.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace MarkovSite.Models
+﻿namespace MarkovSite.Models
 {
     public class GetPredictionsRequest
     {
